+++ src/main/java/com/likelionknu/applyserver/application/data/repository/ApplicationRepository.java
@@ -1,8 +1,8 @@
 package com.likelionknu.applyserver.application.data.repository;
 
+import com.likelionknu.applyserver.auth.data.entity.User;
 import com.likelionknu.applyserver.admin.data.dto.response.AdminRecruitSummaryResponse;
 import com.likelionknu.applyserver.application.data.entity.Application;
-import com.likelionknu.applyserver.auth.data.entity.User;
 import com.likelionknu.applyserver.auth.data.enums.ApplicationStatus;
 import com.likelionknu.applyserver.recruit.data.entity.Recruit;
 import org.springframework.data.jpa.repository.JpaRepository;
